backend/requirements.txt
# Django and related packages
Django==4.2.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-prometheus==2.3.1
sqlparse==0.5.1

# Database related
mysqlclient==2.1.1

# Monitoring and metrics
prometheus_client==0.20.0
psutil==6.0.0

# Utilities and helpers
python-decouple==3.8
requests==2.32.3

# Authentication (if using social auth)
social-auth-app-django==5.4.2
social-auth-core==4.5.4

# Optional - if your project needs it
protobuf==3.12.4
pyparsing==2.4.7
six==1.16.0
typing_extensions==4.12.2
zipp==1.0.0
